requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
boto3==1.35.85
botocore==1.35.99
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
compressed-rtf==1.0.7
cron-descriptor==1.4.5
cryptography==45.0.3
cssselect2==0.8.0
dateparser==1.2.0
defusedxml==0.7.1
Django==4.2.18
django-activity-stream==2.0.0
django-auth-ldap==5.1.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-formtools==2.3
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-model-utils==5.0.0
django-mptt==0.16.0
django-qsstats-magic==1.1.0
django-solo==2.4.0
django-storages==1.14.4
django-stronghold==0.4.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
drf-yasg==1.21.8
easygui==0.98.3
ebcdic==1.1.1
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
extract-msg==0.52.0
flex==6.14.1
furl==2.1.3
fusepy==3.0.1
gevent==24.11.1
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflection==0.5.1
jmespath==1.0.1
josepy==2.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
lark==1.1.9
Markdown==3.7
mozilla-django-oidc==4.0.1
msoffcrypto-tool==5.4.2
nh3==0.2.18
node-semver==0.9.0
olefile==0.47
oletools==0.60.2
ollama==0.4.4
orderedmultidict==1.0.1
packaging==25.0
pcodedmp==1.2.6
pillow==11.0.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.5
pydantic_core==2.33.2
pyotp==2.9.0
pyparsing==3.2.3
pypdf==5.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-gnupg==0.5.3
python-ldap==3.4.4
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
red-black-tree-mod==1.20
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3987==1.3.8
rpds-py==0.25.1
rsa==4.9.1
RTFDE==0.1.2.1
s3transfer==0.10.4
sentry-sdk==2.19.2
setuptools==75.6.0
sh==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
strict-rfc3339==0.7
swagger-spec-validator==3.0.4
tinycss2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
validate_email==1.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.8.2
Whoosh==2.7.4
zipp==3.22.0
zope.event==5.0
zope.interface==7.2
